Only sets FlutterLocalNotificationsPlugin as delegate when initialised
Rather than setting the delegate when the plugin is registered, only set it when the user initialises this plugin.

This solves an issue that arises when this plugin is registered to a background runner but not actually initialised. Previoulsy this would change the UNUserNotificationCenter delegate, and the foreground Flutter instance would no longer receive UNUserNotificationCenter callbacks.